pyproject.toml
[project]
name = "cv-creator-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.124.4",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "PyJWT>=2.10.1",
    "python-multipart>=0.0.20",
    "resend>=2.19.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "fastapi-cli>=0.0.5"
]
